Removed typescript
