+++ package.json
@@ -21,8 +21,6 @@
     "@babel/preset-env": "^7.23.2",
     "@babel/preset-react": "^7.22.15",
     "@babel/preset-typescript": "^7.23.2",
-    "@types/react": "^18.2.33",
-    "@types/react-dom": "^18.2.14",
     "autoprefixer": "^10.4.16",
     "babel-loader": "^9.1.3",
     "babel-plugin-transform-cx-jsx": "^21.3.0",
@@ -34,8 +32,6 @@
     "sass-loader": "^13.3.2",
     "style-loader": "^3.3.3",
     "tailwindcss": "^3.3.5",
-    "ts-loader": "^9.5.0",
-    "typescript": "^5.2.2",
     "webpack": "^5.89.0",
     "webpack-cli": "^5.1.4",
     "webpack-dev-server": "^4.15.1"
