--- tsconfig.json
@@ -1,16 +1,0 @@
-{
-    "compilerOptions": {
-        "outDir": "./dist",
-        "module": "commonjs",
-        "target": "es6",
-        "jsx": "react",
-        "jsxFactory": "cx/src/jsx.js",
-        "jsxFragmentFactory": "cx/src/jsx.js",
-        "strict": true,
-        "moduleResolution": "node",
-        "esModuleInterop": true,
-    },
-    "include": [
-        "src/**/*"
-    ]
-}
